pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "amp-assistant"
version = "0.1.0"
description = "Local, open-source personal assistant with emergent 2-agent collaboration"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
keywords = ["assistant", "ai", "llm", "emergent", "multi-agent", "personal-assistant"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "openai>=1.30.0",
    "anthropic>=0.25.0",
    "python-telegram-bot>=21.0",
    "pyyaml>=6.0",
    "click>=8.1",
    "rich>=13.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "mypy>=1.8",
    "ruff>=0.4",
]

[project.scripts]
amp = "amp.interfaces.cli:cli"

[project.urls]
Homepage = "https://github.com/amp-assistant/amp"
Repository = "https://github.com/amp-assistant/amp"
Issues = "https://github.com/amp-assistant/amp/issues"

[tool.hatch.build.targets.wheel]
packages = ["amp"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.mypy]
python_version = "3.11"
strict = false
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
